+++ packages/network-contracts/contracts/OperatorTokenomics/OperatorPolicies/IDelegationPolicy.sol
@@ -3,7 +3,7 @@
 pragma solidity ^0.8.13;
 
 interface IDelegationPolicy {
-    function setParam(uint256 param) external;
+    function setParam(uint param) external;
 
     /** can throw to prevent delegation. Gets called AFTER the pool token minting */
     function onDelegate(address delegator) external;
